黄祖林提交
2015年8月4日 20:52:49

+++ redpack-web/src/main/java/com/redpack/dao/account/IUserDao.java
@@ -38,7 +38,7 @@ public interface IUserDao {
      * @param userDo
      * @return
      */
-    int saveUser(UserDo userDo);
+    Long saveUser(UserDo userDo);
     /**
      * 根据条件查询用户对象
      * @param parameterMap
